requirements.txt
customtkinter>=5.2.0

# Optional dependencies
Pillow>=10.0.0
